Moved AppEngineComponentPage to c.g.c.t.e.appengine.ui
