+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/META-INF/MANIFEST.MF
@@ -7,5 +7,15 @@ Bundle-Vendor: Google Inc.
 Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: com.google.cloud.tools.eclipse.appengine.ui
-Import-Package: org.eclipse.jface.resource,
+Import-Package: org.eclipse.jface.dialogs,
+ org.eclipse.jface.layout,
+ org.eclipse.jface.resource,
+ org.eclipse.jface.wizard,
+ org.eclipse.swt,
+ org.eclipse.swt.custom,
+ org.eclipse.swt.graphics,
+ org.eclipse.swt.layout,
+ org.eclipse.swt.widgets,
  org.eclipse.ui.plugin
+Require-Bundle: com.google.cloud.tools.eclipse.usagetracker,
+ org.eclipse.osgi
