+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/src/com/google/cloud/tools/eclipse/appengine/ui/messages.properties
@@ -1,0 +1,1 @@
+AppEngineJavaComponentMissing=Cannot create an App Engine Eclipse project because the Cloud SDK App Engine Java component is not installed. Fix by running 'gcloud components install app-engine-java' on the command-line.
